LayoutTests/editing/selection/ios/select-text-in-existing-selection-expected.txt

The quick brown fox jumped over the lazy dog.
This test verifies that the selection can be modified via long press inside an existing selection. To manually test, select all the text in the editable area and long press on a word to select it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().toString() is "jumped"
PASS successfullyParsed is true

TEST COMPLETE
